+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][],
 and this project adheres to [Semantic Versioning][].
 
-## 1.0.0-beta2
+## [1.0.0-beta2]
 
 ## Added
 
@@ -24,3 +24,4 @@ Initial beta release
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 [1.0.0-beta1]: https://github.com/mishamyrt/vpn-dns/releases/tag/v1.0.0-beta1
+[1.0.0-beta2]: https://github.com/mishamyrt/vpn-dns/releases/tag/v1.0.0-beta2